+++ first_app/lib/gradient_container.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/material.dart';
+
+class GradientContainer extends StatelessWidget {
+  // initialization work
+  const GradientContainer({super.key});
+
+  @override
+  Widget build(context) {
+    return Container(
+      decoration: const BoxDecoration(
+        gradient: LinearGradient(
+          colors: [
+            Colors.amber,
+            Colors.deepPurple,
+          ],
+          begin: Alignment.topLeft,
+          end: Alignment.bottomRight,
+        ),
+      ),
+      child: const Center(
+        child: Text(
+          'Hello World!',
+          style: TextStyle(
+            color: Colors.white,
+            fontSize: 20,
+          ),
+        ),
+      ),
+    );
+  }
+}
