+++ first_app/lib/main.dart
@@ -1,43 +1,16 @@
 import 'package:flutter/material.dart';
 
+import 'package:first_app/gradient_container.dart';
+
 void main() {
   runApp(
-    MaterialApp(
+    const MaterialApp(
       home: Scaffold(
         backgroundColor: Colors.deepPurple,
         body: GradientContainer(),
       ),
     ),
   );
-}
-
-
-class GradientContainer extends StatelessWidget {
-  @override
-  Widget build(context) {
-    return 
-    Container(
-          decoration: const BoxDecoration(
-            gradient: LinearGradient(
-              colors: [
-                Colors.amber,
-                Colors.deepPurple,
-              ],
-              begin: Alignment.topLeft,
-              end: Alignment.bottomRight,
-            ),
-          ),
-          child: const Center(
-            child: Text(
-              'Hello World!',
-              style: TextStyle(
-                color: Colors.white,
-                fontSize: 20,
-              ),
-            ),
-          ),
-        );
-  }
 }
